Refactoring, tests, and cleanup
