+++ enforcer/utils/tokens/jwt_test.go
@@ -173,81 +173,3 @@ func TestCreateAndVerifyPKI(t *testing.T) {
 
 	})
 }
-
-//
-//
-// func TestCreateAndVerifyPKI(t *testing.T) {
-// 	Convey("Given  a new JWT Engine with PKI and inband certificates", t, func() {
-//
-// 		jwtConfig := NewPKIJWT(validity, "TRIREME", nil, []byte(keyPEM), []byte(certPEM), []byte(caPool))
-//
-// 		Convey("Given a signature request for a normal packet ", func() {
-// 			token := jwtConfig.CreateAndSign(false, &defaultClaims)
-// 			recoveredClaims, cert := jwtConfig.Decode(false, token, nil)
-//
-// 			So(recoveredClaims, ShouldNotBeNil)
-// 			So(recoveredClaims.T["label1"], ShouldEqual, defaultClaims.T["label1"])
-// 			So(recoveredClaims.T["label2"], ShouldEqual, defaultClaims.T["label2"])
-// 			So(string(recoveredClaims.RMT), ShouldEqual, rmt)
-// 			So(string(recoveredClaims.LCL), ShouldEqual, lcl)
-// 			So(cert, ShouldNotBeNil)
-//
-// 		})
-//
-// 		Convey("Given a singature request for an ACK packet, where there is no sign certificate", func() {
-// 			token := jwtConfig.CreateAndSign(true, &ackClaims)
-// 			recoveredClaims, _ := jwtConfig.Decode(true, token, nil)
-//
-// 			So(recoveredClaims, ShouldBeNil)
-// 		})
-//
-// 		Convey("Given a signature request for an ack packet where the right certificate is provided", func() {
-// 			_, cert, _, _ := loadKeys([]byte(keyPEM), []byte(certPEM), []byte(caPool))
-// 			token := jwtConfig.CreateAndSign(true, &ackClaims)
-// 			recoveredClaims, _ := jwtConfig.Decode(true, token, cert)
-// 			So(recoveredClaims, ShouldNotBeNil)
-// 		})
-// 	})
-//
-// 	Convey("Given a token engine with out-of-band certificate generation", t, func() {
-// 		_, cert, _, _ := loadKeys([]byte(keyPEM), []byte(certPEM), []byte(caPool))
-//
-// 		certCache := map[string]*ecdsa.PublicKey{
-// 			"TRIREME": cert.PublicKey.(*ecdsa.PublicKey),
-// 		}
-//
-// 		jwtConfig := NewPKIJWT(validity, "TRIREME", certCache, []byte(keyPEM), []byte(certPEM), []byte(caPool))
-// 		So(jwtConfig.IncludeCert, ShouldBeFalse)
-// 		So(jwtConfig.PublicKeyCache, ShouldEqual, certCache)
-// 		So(jwtConfig.PublicKeyCache["TRIREME"], ShouldEqual, cert.PublicKey.(*ecdsa.PublicKey))
-//
-// 		Convey("Given a signature request for a normal packet ", func() {
-// 			token := jwtConfig.CreateAndSign(false, &defaultClaims)
-// 			recoveredClaims, _ := jwtConfig.Decode(false, token, nil)
-//
-// 			So(recoveredClaims, ShouldNotBeNil)
-// 			So(recoveredClaims.T["label1"], ShouldEqual, defaultClaims.T["label1"])
-// 			So(recoveredClaims.T["label2"], ShouldEqual, defaultClaims.T["label2"])
-// 			So(string(recoveredClaims.RMT), ShouldEqual, rmt)
-// 			So(string(recoveredClaims.LCL), ShouldEqual, lcl)
-//
-// 		})
-//
-// 		Convey("Given a signature request for a normal packet, where the cache doesn't have the certificate ", func() {
-// 			token := jwtConfig.CreateAndSign(false, &defaultClaims)
-// 			jwtConfig.PublicKeyCache = nil
-// 			recoveredClaims, _ := jwtConfig.Decode(false, token, nil)
-//
-// 			So(recoveredClaims, ShouldBeNil)
-//
-// 		})
-//
-// 		Convey("Given a signature request for an ack packet where the right certificate is provided", func() {
-// 			token := jwtConfig.CreateAndSign(true, &ackClaims)
-// 			recoveredClaims, _ := jwtConfig.Decode(true, token, nil)
-//
-// 			So(recoveredClaims, ShouldNotBeNil)
-// 		})
-// 	})
-//
-// }
